requirements.txt
django>=5.0,<6.0
django-extensions
psycopg2-binary
django-storages
boto3
google-cloud-storage
celery[redis]
redis
django-otp
django-two-factor-auth
djangorestframework
